requirements.txt
streamlit
pandas
numpy
seaborn
statsmodels
plotly
matplotlib
python-dateutil